+++ config-ui/Common/src/main/java/de/markusbordihn/easynpc/configui/client/renderer/manager/EntityTypeValidator.java
@@ -51,7 +51,7 @@ public class EntityTypeValidator {
       }
 
       if (entityType != null) {
-        EntityTypeManager.getPathfinderMob(entityType, level);
+        EntityTypeManager.getRenderEntity(entityType, level);
         processed++;
       }
     }