+++ core/Common/src/main/java/de/markusbordihn/easynpc/data/state/StateOperation.java
@@ -34,6 +34,7 @@ public enum StateOperation {
 
   private static final int DEFAULT_AMOUNT = 1;
   private static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
+  private final String commandName = this.name().toLowerCase(Locale.ROOT);
 
   public static StateOperation get(String stateOperation) {
     return EnumUtils.getIgnoreCase(StateOperation.class, stateOperation, SET);
@@ -65,7 +66,7 @@ public enum StateOperation {
   }
 
   public String getCommandName() {
-    return this.name().toLowerCase(Locale.ROOT);
+    return this.commandName;
   }
 
   public boolean requiresValue() {